Add @new directive
Methods can be added to classes at runtime via the @new directive. It is used is the same way as @optional and @reqired, but is only valid within class interfaces and categories.
If a method within an @hook group is specified as @new, it is added to the hooked class at runtime.