Use EmitAnyExpr to emit arguments to @orig
EmitCallArgs does not function correctly for what we're trying to do. If argument expressions are lvalues, EmitCallArg (which is called by EmitCallArgs) emits the *address* of the passed argument, not the argument itself.

e.g, @orig(some_lvalue) would actually evaluate @orig(&some_lvalue)

EmitAnyExpr emits expressions as an rvalue, which is suitable for our needs.